refactor core long-file modules
