+++ agently/core/orchestration/TriggerFlow/BluePrint.py
@@ -17,11 +17,8 @@ import uuid
 import copy
 import json
 import asyncio
-import contextlib
-import re
 import hashlib
 import yaml
-from dataclasses import dataclass
 from pathlib import Path
 from json import JSONDecodeError
 from asyncio import Event, Semaphore
@@ -41,13 +38,10 @@ if TYPE_CHECKING:
     from agently.types.data import ExecutionEnvironmentRequirement
     from .TriggerFlow import TriggerFlow
 
-from agently.types.data import EMPTY, SerializableMapping
-from agently.types.trigger_flow import RUNTIME_STREAM_STOP
-from agently.utils import StateData, StateDataNamespace
-from agently.core.runtime.RuntimeContext import resolve_parent_run_context
+from agently.types.data import EMPTY
+from agently.utils import StateDataNamespace
 from .Chunk import TriggerFlowChunk
 from .Execution import TriggerFlowExecution
-from .Control import TriggerFlowPauseSignal, TRIGGER_FLOW_STATUS_WAITING
 from .Definition import (
     TriggerFlowDefinition,
     build_callable_ref,
@@ -55,154 +49,12 @@ from .Definition import (
     make_signal_ref,
     render_callable_ref,
 )
-
-_SUB_FLOW_PATH_SEGMENT_PATTERN = re.compile(r"^[A-Za-z_][A-Za-z0-9_]*$")
-_CAPTURE_TARGET_SCOPES = frozenset({"input", "runtime_data", "flow_data", "resources"})
-_CAPTURE_SOURCE_SCOPES = frozenset({"value", "runtime_data", "flow_data", "resources"})
-_WRITE_BACK_TARGET_SCOPES = frozenset({"value", "runtime_data", "flow_data"})
-_WRITE_BACK_SOURCE_SCOPES = frozenset({"result"})
+from .BluePrintSubFlow import TriggerFlowBlueprintSubFlow
+from .SubFlowBindings import _CompiledSubFlowBinding
 
 
 def _stable_definition_json(value: Any):
     return json.dumps(value, sort_keys=True, separators=(",", ":"), default=repr)
-
-
-@dataclass(frozen=True)
-class _CompiledSubFlowBinding:
-    target_scope: Literal["input", "runtime_data", "flow_data", "resources", "value"]
-    target_path: tuple[str, ...]
-    source_scope: Literal["value", "runtime_data", "flow_data", "resources", "result"]
-    source_path: tuple[str, ...]
-
-
-def _clone_sub_flow_value(value: Any):
-    try:
-        return copy.deepcopy(value)
-    except Exception:
-        return value
-
-
-def _read_sub_flow_value_by_path(root_value: Any, path: tuple[str, ...], *, scope: str):
-    current = root_value
-    for segment in path:
-        if isinstance(current, dict) and segment in current:
-            current = current[segment]
-            continue
-        raise KeyError(f"TriggerFlow sub flow path '{ scope }.{ '.'.join(path) }' not found.")
-    return _clone_sub_flow_value(current)
-
-
-class _ParentSubFlowCaptureSource:
-    def __init__(self, data):
-        flow_data = data.get_flow_data(None, {}, no_warning=True)
-        self._scopes = {
-            "value": data.value,
-            "runtime_data": data.state.to_dict(),
-            "flow_data": flow_data if isinstance(flow_data, dict) else {},
-            "resources": data.resources.to_dict(),
-        }
-
-    def read_path(self, scope: str, path: tuple[str, ...]):
-        return _read_sub_flow_value_by_path(self._scopes[scope], path, scope=scope)
-
-
-class _SubFlowWriteBackSource:
-    def __init__(self, result: Any):
-        self._result = result
-
-    def read_path(self, scope: str, path: tuple[str, ...]):
-        if scope != "result":
-            raise KeyError(f"Unsupported TriggerFlow sub flow write back source scope '{ scope }'.")
-        if isinstance(self._result, dict) and "$final_result" in self._result:
-            try:
-                return _read_sub_flow_value_by_path(self._result["$final_result"], path, scope=scope)
-            except KeyError:
-                return _read_sub_flow_value_by_path(self._result, path, scope=scope)
-        return _read_sub_flow_value_by_path(self._result, path, scope=scope)
-
-
-class _SubFlowCaptureTarget:
-    def __init__(self):
-        self._has_input = False
-        self._input_value = None
-        self._runtime_data = StateData()
-        self._flow_data = StateData()
-        self._resources = StateData()
-
-    def write_path(self, scope: str, path: tuple[str, ...], value: Any):
-        copied_value = _clone_sub_flow_value(value)
-        if scope == "input":
-            if len(path) == 0:
-                self._input_value = copied_value
-                self._has_input = True
-                return
-            current_input = self._input_value if isinstance(self._input_value, dict) else {}
-            input_data = StateData(_clone_sub_flow_value(current_input))
-            input_data.set(".".join(path), copied_value)
-            self._input_value = input_data.get(None, {}, inherit=False)
-            self._has_input = True
-            return
-
-        target_scope = {
-            "runtime_data": self._runtime_data,
-            "flow_data": self._flow_data,
-            "resources": self._resources,
-        }[scope]
-        target_scope.set(".".join(path), copied_value)
-
-    def build_input(self):
-        return _clone_sub_flow_value(self._input_value) if self._has_input else None
-
-    def build_runtime_data(self):
-        return self._runtime_data.get(None, {}, inherit=False)
-
-    def build_flow_data(self):
-        return self._flow_data.get(None, {}, inherit=False)
-
-    def build_resources(self):
-        return self._resources.get(None, {}, inherit=False)
-
-
-class _SubFlowWriteBackTarget:
-    def __init__(self, initial_value: Any):
-        self._has_value_binding = False
-        self._current_value = _clone_sub_flow_value(initial_value)
-        initial_mapping = self._current_value if isinstance(self._current_value, dict) else {}
-        self._value_data = StateData(_clone_sub_flow_value(initial_mapping))
-        self._runtime_data = StateData()
-        self._flow_data = StateData()
-
-    def write_path(self, scope: str, path: tuple[str, ...], value: Any):
-        copied_value = _clone_sub_flow_value(value)
-        if scope == "value":
-            self._has_value_binding = True
-            if len(path) == 0:
-                self._current_value = copied_value
-                return
-            if not isinstance(self._current_value, dict):
-                self._current_value = {}
-                self._value_data = StateData({})
-            self._value_data.set(".".join(path), copied_value)
-            self._current_value = self._value_data.get(None, {}, inherit=False)
-            return
-
-        target_scope = {
-            "runtime_data": self._runtime_data,
-            "flow_data": self._flow_data,
-        }[scope]
-        target_scope.set(".".join(path), copied_value)
-
-    def apply(self, data):
-        if self._has_value_binding:
-            data.value = _clone_sub_flow_value(self._current_value)
-
-        runtime_data_patch = self._runtime_data.get(None, {}, inherit=False)
-        for key, value in runtime_data_patch.items():
-            data.execution._runtime_data.set(key, _clone_sub_flow_value(value))
-
-        flow_data_patch = self._flow_data.get(None, {}, inherit=False)
-        for key, value in flow_data_patch.items():
-            data.execution._trigger_flow._flow_data.set(key, _clone_sub_flow_value(value))
 
 
 class TriggerFlowBlueprint:
@@ -217,6 +69,7 @@ class TriggerFlowBlueprint:
         self.definition = TriggerFlowDefinition(name=self.name)
         self._chunk_registry: dict[str, Any] = {}
         self._condition_registry: dict[str, Any] = {}
+        self._sub_flow = TriggerFlowBlueprintSubFlow(self)
 
     def make_stable_identity_digest(self, identity: Any):
         digest = hashlib.sha1(_stable_definition_json(identity).encode("utf-8")).hexdigest()[:16]
@@ -477,19 +330,7 @@ class TriggerFlowBlueprint:
         return self.definition.get_operator(chunk.id)
 
     def _parse_sub_flow_relative_path(self, path: str, *, option_name: str):
-        if not isinstance(path, str):
-            raise TypeError(
-                f"TriggerFlow sub flow { option_name } target path must be a string, got: { type(path) }."
-            )
-        if path == "":
-            raise ValueError(f"TriggerFlow sub flow { option_name } target path can not be empty.")
-        segments = tuple(path.split("."))
-        for segment in segments:
-            if not _SUB_FLOW_PATH_SEGMENT_PATTERN.fullmatch(segment):
-                raise ValueError(
-                    f"TriggerFlow sub flow { option_name } target path '{ path }' contains invalid segment '{ segment }'."
-                )
-        return segments
+        return self._sub_flow.parse_relative_path(path, option_name=option_name)
 
     def _parse_sub_flow_source_path(
         self,
@@ -498,29 +339,11 @@ class TriggerFlowBlueprint:
         mode: Literal["capture", "write_back"],
         target_scope: str,
     ):
-        if not isinstance(path, str):
-            raise TypeError(
-                f"TriggerFlow sub flow { mode } source path for target scope '{ target_scope }' "
-                f"must be a string, got: { type(path) }."
-            )
-        if path == "":
-            raise ValueError(
-                f"TriggerFlow sub flow { mode } source path for target scope '{ target_scope }' can not be empty."
-            )
-        segments = tuple(path.split("."))
-        root_scope = segments[0]
-        allowed_source_scopes = _CAPTURE_SOURCE_SCOPES if mode == "capture" else _WRITE_BACK_SOURCE_SCOPES
-        if root_scope not in allowed_source_scopes:
-            raise ValueError(
-                f"TriggerFlow sub flow { mode } source scope '{ root_scope }' is not supported. "
-                f"Allowed scopes: { sorted(allowed_source_scopes) }"
-            )
-        for segment in segments[1:]:
-            if not _SUB_FLOW_PATH_SEGMENT_PATTERN.fullmatch(segment):
-                raise ValueError(
-                    f"TriggerFlow sub flow { mode } source path '{ path }' contains invalid segment '{ segment }'."
-                )
-        return root_scope, segments[1:]
+        return self._sub_flow.parse_source_path(
+            path,
+            mode=mode,
+            target_scope=target_scope,
+        )
 
     def _normalize_sub_flow_scope_binding(
         self,
@@ -529,32 +352,11 @@ class TriggerFlowBlueprint:
         mode: Literal["capture", "write_back"],
         scope: str,
     ):
-        if isinstance(binding, str):
-            if scope not in {"input", "value"}:
-                raise TypeError(
-                    f"TriggerFlow sub flow { mode } scope '{ scope }' only accepts key-path mappings."
-                )
-            return binding
-
-        if not isinstance(binding, Mapping):
-            raise TypeError(
-                f"TriggerFlow sub flow { mode } scope '{ scope }' expects a string or mapping, got: { type(binding) }."
-            )
-
-        normalized_binding: dict[str, str] = {}
-        for target_path, source_path in binding.items():
-            if not isinstance(target_path, str):
-                raise TypeError(
-                    f"TriggerFlow sub flow { mode } target path for scope '{ scope }' must be a string, "
-                    f"got: { type(target_path) }."
-                )
-            if not isinstance(source_path, str):
-                raise TypeError(
-                    f"TriggerFlow sub flow { mode } source path for scope '{ scope }' must be a string, "
-                    f"got: { type(source_path) }."
-                )
-            normalized_binding[str(target_path)] = str(source_path)
-        return normalized_binding
+        return self._sub_flow.normalize_scope_binding(
+            binding,
+            mode=mode,
+            scope=scope,
+        )
 
     def _normalize_sub_flow_spec(
         self,
@@ -562,29 +364,7 @@ class TriggerFlowBlueprint:
         *,
         mode: Literal["capture", "write_back"],
     ):
-        if spec is None:
-            return None
-        if not isinstance(spec, Mapping):
-            raise TypeError(f"TriggerFlow sub flow { mode } spec must be a mapping, got: { type(spec) }.")
-
-        allowed_target_scopes = _CAPTURE_TARGET_SCOPES if mode == "capture" else _WRITE_BACK_TARGET_SCOPES
-        normalized_spec: dict[str, str | dict[str, str]] = {}
-        for target_scope, binding in spec.items():
-            if not isinstance(target_scope, str):
-                raise TypeError(
-                    f"TriggerFlow sub flow { mode } target scope must be a string, got: { type(target_scope) }."
-                )
-            if target_scope not in allowed_target_scopes:
-                raise ValueError(
-                    f"TriggerFlow sub flow { mode } target scope '{ target_scope }' is not supported. "
-                    f"Allowed scopes: { sorted(allowed_target_scopes) }"
-                )
-            normalized_spec[target_scope] = self._normalize_sub_flow_scope_binding(
-                binding,
-                mode=mode,
-                scope=target_scope,
-            )
-        return normalized_spec
+        return self._sub_flow.normalize_spec(spec, mode=mode)
 
     def _validate_sub_flow_target_conflicts(
         self,
@@ -593,19 +373,11 @@ class TriggerFlowBlueprint:
         mode: Literal["capture", "write_back"],
         scope: str,
     ):
-        seen_paths: set[tuple[str, ...]] = set()
-        for target_path in sorted(target_paths, key=len):
-            if target_path in seen_paths:
-                raise ValueError(
-                    f"TriggerFlow sub flow { mode } target path '{ scope }.{ '.'.join(target_path) }' is duplicated."
-                )
-            for depth in range(1, len(target_path)):
-                if target_path[:depth] in seen_paths:
-                    raise ValueError(
-                        f"TriggerFlow sub flow { mode } target paths conflict under scope '{ scope }': "
-                        f"'{ scope }.{ '.'.join(target_path[:depth]) }' and '{ scope }.{ '.'.join(target_path) }'."
-                    )
-            seen_paths.add(target_path)
+        return self._sub_flow.validate_target_conflicts(
+            target_paths,
+            mode=mode,
+            scope=scope,
+        )
 
     def _compile_sub_flow_bindings(
         self,
@@ -613,86 +385,7 @@ class TriggerFlowBlueprint:
         *,
         mode: Literal["capture", "write_back"],
     ):
-        normalized_spec = self._normalize_sub_flow_spec(spec, mode=mode)
-        compiled_bindings: list[_CompiledSubFlowBinding] = []
-
-        if normalized_spec is None:
-            default_target_scope = "input" if mode == "capture" else "value"
-            default_source_path = "value" if mode == "capture" else "result"
-            source_scope, source_path = self._parse_sub_flow_source_path(
-                default_source_path,
-                mode=mode,
-                target_scope=default_target_scope,
-            )
-            compiled_bindings.append(
-                _CompiledSubFlowBinding(
-                    target_scope=default_target_scope,
-                    target_path=tuple(),
-                    source_scope=cast(
-                        Literal["value", "runtime_data", "flow_data", "resources", "result"],
-                        source_scope,
-                    ),
-                    source_path=tuple(source_path),
-                )
-            )
-            return normalized_spec, compiled_bindings
-
-        for target_scope, binding in normalized_spec.items():
-            if isinstance(binding, str):
-                source_scope, source_path = self._parse_sub_flow_source_path(
-                    binding,
-                    mode=mode,
-                    target_scope=target_scope,
-                )
-                compiled_bindings.append(
-                    _CompiledSubFlowBinding(
-                        target_scope=cast(
-                            Literal["input", "runtime_data", "flow_data", "resources", "value"],
-                            target_scope,
-                        ),
-                        target_path=tuple(),
-                        source_scope=cast(
-                            Literal["value", "runtime_data", "flow_data", "resources", "result"],
-                            source_scope,
-                        ),
-                        source_path=tuple(source_path),
-                    )
-                )
-                continue
-
-            target_paths: list[tuple[str, ...]] = []
-            for target_path_str, source_path_str in binding.items():
-                target_path = self._parse_sub_flow_relative_path(
-                    target_path_str,
-                    option_name=f"{ mode }.{ target_scope }",
-                )
-                source_scope, source_path = self._parse_sub_flow_source_path(
-                    source_path_str,
-                    mode=mode,
-                    target_scope=target_scope,
-                )
-                target_paths.append(target_path)
-                compiled_bindings.append(
-                    _CompiledSubFlowBinding(
-                        target_scope=cast(
-                            Literal["input", "runtime_data", "flow_data", "resources", "value"],
-                            target_scope,
-                        ),
-                        target_path=target_path,
-                        source_scope=cast(
-                            Literal["value", "runtime_data", "flow_data", "resources", "result"],
-                            source_scope,
-                        ),
-                        source_path=tuple(source_path),
-                    )
-                )
-            self._validate_sub_flow_target_conflicts(
-                target_paths,
-                mode=mode,
-                scope=target_scope,
-            )
-
-        return normalized_spec, compiled_bindings
+        return self._sub_flow.compile_bindings(spec, mode=mode)
 
     def _apply_sub_flow_bindings(
         self,
@@ -701,56 +394,23 @@ class TriggerFlowBlueprint:
         source: "TriggerFlowPathReadable",
         target: "TriggerFlowPathWritable",
     ):
-        for binding in bindings:
-            value = source.read_path(binding.source_scope, binding.source_path)
-            target.write_path(binding.target_scope, binding.target_path, value)
+        return self._sub_flow.apply_bindings(bindings, source=source, target=target)
 
     def _instantiate_isolated_sub_flow(self, trigger_flow: "TriggerFlow"):
-        from .TriggerFlow import TriggerFlow
-
-        isolated_sub_flow = TriggerFlow(
-            blueprint=trigger_flow.save_blueprint(),
-            name=trigger_flow.name,
-            skip_exceptions=trigger_flow._skip_exceptions,
-        )
-        settings_snapshot = copy.deepcopy(trigger_flow.settings.get(None, {}, inherit=False))
-        if not isinstance(settings_snapshot, Mapping):
-            raise TypeError(
-                f"TriggerFlow settings snapshot must be a mapping, got: { type(settings_snapshot) }."
-            )
-        isolated_sub_flow.settings.update(cast(SerializableMapping, settings_snapshot))
-        isolated_sub_flow._flow_data.update(
-            copy.deepcopy(trigger_flow._flow_data.get(None, {}, inherit=False))
-        )
-        isolated_sub_flow._runtime_resources.update(
-            copy.deepcopy(trigger_flow._runtime_resources.get(None, {}, inherit=False))
-        )
-        return isolated_sub_flow
+        return self._sub_flow.instantiate_isolated_sub_flow(trigger_flow)
 
     async def _bridge_sub_flow_runtime_stream(
         self,
         child_execution: TriggerFlowExecution,
         parent_execution: TriggerFlowExecution,
     ):
-        while True:
-            stream_item = await child_execution._runtime_stream_queue.get()
-            if stream_item is RUNTIME_STREAM_STOP:
-                return
-            await parent_execution.async_put_into_stream(stream_item)
+        return await self._sub_flow.bridge_runtime_stream(child_execution, parent_execution)
 
     def _build_sub_flow_resource_bindings(
         self,
         capture_bindings: Sequence[_CompiledSubFlowBinding],
     ):
-        bindings: dict[str, str] = {}
-        for binding in capture_bindings:
-            if binding.target_scope != "resources" or binding.source_scope != "resources":
-                continue
-            target_key = ".".join(binding.target_path)
-            source_key = ".".join(binding.source_path)
-            if target_key and source_key:
-                bindings[target_key] = source_key
-        return bindings
+        return self._sub_flow.build_resource_bindings(capture_bindings)
 
     def _apply_sub_flow_resource_bindings(
         self,
@@ -758,11 +418,11 @@ class TriggerFlowBlueprint:
         parent_execution: TriggerFlowExecution,
         resource_bindings: Mapping[str, str],
     ):
-        for target_key, source_key in resource_bindings.items():
-            child_execution.set_runtime_resource(
-                str(target_key),
-                parent_execution.require_runtime_resource(str(source_key)),
-            )
+        return self._sub_flow.apply_resource_bindings(
+            child_execution,
+            parent_execution,
+            resource_bindings,
+        )
 
     def _restore_sub_flow_frame_resources(
         self,
@@ -770,24 +430,10 @@ class TriggerFlowBlueprint:
         parent_execution: TriggerFlowExecution,
         frame: dict[str, Any],
     ):
-        resource_bindings = frame.get("resource_bindings", {})
-        if isinstance(resource_bindings, Mapping):
-            self._apply_sub_flow_resource_bindings(
-                child_execution,
-                parent_execution,
-                resource_bindings,
-            )
-        for resource_key in frame.get("resource_keys", []):
-            key = str(resource_key)
-            if child_execution.get_runtime_resource(key, EMPTY) is not EMPTY:
-                continue
-            child_execution.set_runtime_resource(
-                key,
-                parent_execution.require_runtime_resource(key),
-            )
+        return self._sub_flow.restore_frame_resources(child_execution, parent_execution, frame)
 
     def _make_sub_flow_frame_id(self, parent_execution: TriggerFlowExecution, operator_id: str):
-        return f"{ parent_execution.id }:{ operator_id }:{ uuid.uuid4().hex }"
+        return self._sub_flow.make_frame_id(parent_execution, operator_id)
 
     def _build_sub_flow_parent_data(
         self,
@@ -795,26 +441,7 @@ class TriggerFlowBlueprint:
         frame: dict[str, Any],
         operator: dict[str, Any],
     ):
-        from agently.types.trigger_flow import TriggerFlowRuntimeData
-
-        parent_signal = parent_execution._restore_signal(frame.get("parent_signal"))
-        if parent_signal is None:
-            parent_signal = parent_execution._build_signal(
-                "START",
-                frame.get("parent_value"),
-                trigger_type="event",
-                source="sub_flow",
-            )
-        chunk_run_context = parent_execution._create_chunk_run_context(operator, parent_signal)
-        return TriggerFlowRuntimeData(
-            trigger_event=parent_signal.trigger_event,
-            trigger_type=parent_signal.trigger_type,
-            value=frame.get("parent_value"),
-            execution=parent_execution,
-            _layer_marks=list(frame.get("parent_layer_marks", [])),
-            signal=parent_signal,
-            chunk_run_context=chunk_run_context,
-        )
+        return self._sub_flow.build_parent_data(parent_execution, frame, operator)
 
     async def _project_child_interrupts(
         self,
@@ -823,56 +450,11 @@ class TriggerFlowBlueprint:
         child_execution: TriggerFlowExecution,
         frame: dict[str, Any],
     ):
-        pending_child_interrupts = child_execution.get_pending_interrupts()
-        projected_interrupts = dict(frame.get("projected_interrupts", {}))
-        parent_interrupts = parent_execution._get_interrupts().copy()
-        projected_root_ids: list[str] = []
-
-        for child_interrupt_id, child_interrupt in pending_child_interrupts.items():
-            root_interrupt_id = None
-            for stored_root_id, stored_child_id in projected_interrupts.items():
-                if stored_child_id == child_interrupt_id:
-                    root_interrupt_id = stored_root_id
-                    break
-            if root_interrupt_id is None:
-                root_interrupt_id = f"{ frame['frame_id'] }:{ child_interrupt_id }"
-                projected_interrupts[root_interrupt_id] = child_interrupt_id
-
-            root_interrupt = copy.deepcopy(child_interrupt)
-            root_interrupt.update(
-                {
-                    "id": root_interrupt_id,
-                    "local_interrupt_id": child_interrupt_id,
-                    "status": "waiting",
-                    "source_execution_id": child_execution.id,
-                    "source_flow_name": child_execution._trigger_flow.name,
-                    "sub_flow_frame_id": frame["frame_id"],
-                    "child_interrupt": copy.deepcopy(child_interrupt),
-                }
-            )
-            parent_interrupts[root_interrupt_id] = root_interrupt
-            projected_root_ids.append(root_interrupt_id)
-
-            await parent_execution.async_put_into_stream(
-                {
-                    "type": "interrupt",
-                    "action": "project",
-                    "execution_id": parent_execution.id,
-                    "interrupt": parent_execution._to_serializable_value(root_interrupt),
-                    "value": parent_execution._to_serializable_value(child_interrupt.get("payload")),
-                },
-                _skip_contract_validation=True,
-            )
-
-        frame["status"] = "waiting"
-        frame["child_saved_state"] = child_execution.save()
-        frame["projected_interrupts"] = projected_interrupts
-        frame["resource_keys"] = sorted(str(key) for key in child_execution.get_runtime_resources().keys())
-        frame["projected_root_interrupt_ids"] = projected_root_ids
-        parent_execution._system_runtime_data.set("interrupts", parent_interrupts)
-        parent_execution._set_sub_flow_frame(frame["frame_id"], frame)
-        parent_execution._set_status(TRIGGER_FLOW_STATUS_WAITING)
-        return projected_root_ids
+        return await self._sub_flow.project_child_interrupts(
+            parent_execution=parent_execution,
+            child_execution=child_execution,
+            frame=frame,
+        )
 
     async def _complete_sub_flow_frame(
         self,
@@ -884,61 +466,17 @@ class TriggerFlowBlueprint:
         normalized_write_back: Any,
         write_back_bindings: Sequence[_CompiledSubFlowBinding],
     ):
-        result = await child_execution.async_close(reason="sub_flow_completed")
-        data = self._build_sub_flow_parent_data(parent_execution, frame, operator)
-        if normalized_write_back is None:
-            if isinstance(result, dict) and "$final_result" in result:
-                data.value = result["$final_result"]
-            else:
-                data.value = result
-        else:
-            write_back_target = _SubFlowWriteBackTarget(data.value)
-            self._apply_sub_flow_bindings(
-                write_back_bindings,
-                source=_SubFlowWriteBackSource(result),
-                target=write_back_target,
-            )
-            write_back_target.apply(data)
-
-        frame["status"] = "completed"
-        frame["child_saved_state"] = child_execution.save()
-        frame["result"] = parent_execution._to_serializable_value(result)
-        parent_execution._set_sub_flow_frame(frame["frame_id"], frame)
-
-        emit_signal = operator["emit_signals"][0]
-        await data.async_emit(
-            emit_signal["trigger_event"],
-            data.value,
-            _layer_marks=data._layer_marks.copy(),
+        return await self._sub_flow.complete_frame(
+            parent_execution=parent_execution,
+            child_execution=child_execution,
+            frame=frame,
+            operator=operator,
+            normalized_write_back=normalized_write_back,
+            write_back_bindings=write_back_bindings,
         )
-        parent_execution._refresh_waiting_status()
-        return result
 
     def _build_sub_flow_from_operator(self, operator: dict[str, Any]):
-        from .TriggerFlow import TriggerFlow
-
-        options = operator.get("options", {})
-        sub_flow_config = options.get("sub_flow_config")
-        if not isinstance(sub_flow_config, dict):
-            raise TypeError(
-                f"TriggerFlow sub flow operator '{ operator['id'] }' missing valid 'sub_flow_config'."
-            )
-
-        sub_blueprint = type(self)(
-            name=str(
-                sub_flow_config.get("name")
-                or options.get("sub_flow_name")
-                or operator.get("name")
-                or f"SubFlow-{ operator['id'] }"
-            )
-        )
-        sub_blueprint._chunk_registry = self._chunk_registry.copy()
-        sub_blueprint._condition_registry = self._condition_registry.copy()
-        sub_blueprint.load_flow_config(copy.deepcopy(sub_flow_config))
-        return TriggerFlow(
-            blueprint=sub_blueprint,
-            name=sub_blueprint.name,
-        )
+        return self._sub_flow.build_from_operator(operator)
 
     def _compile_sub_flow_operator(
         self,
@@ -946,129 +484,7 @@ class TriggerFlowBlueprint:
         *,
         trigger_flow: "TriggerFlow | None" = None,
     ):
-        options = operator.get("options", {})
-        _, capture_bindings = self._compile_sub_flow_bindings(
-            options.get("capture"),
-            mode="capture",
-        )
-        normalized_write_back, write_back_bindings = self._compile_sub_flow_bindings(
-            options.get("write_back"),
-            mode="write_back",
-        )
-        concurrency = operator["options"].get("concurrency")
-        sub_flow_template = trigger_flow if trigger_flow is not None else self._build_sub_flow_from_operator(operator)
-        resource_bindings = self._build_sub_flow_resource_bindings(capture_bindings)
-
-        async def call_sub_flow(data):
-            isolated_sub_flow = self._instantiate_isolated_sub_flow(sub_flow_template)
-
-            capture_source = _ParentSubFlowCaptureSource(data)
-            capture_target = _SubFlowCaptureTarget()
-            self._apply_sub_flow_bindings(
-                capture_bindings,
-                source=capture_source,
-                target=capture_target,
-            )
-
-            captured_flow_data = capture_target.build_flow_data()
-            if captured_flow_data:
-                isolated_sub_flow._flow_data.update(captured_flow_data)
-
-            sub_flow_execution = isolated_sub_flow.create_execution(
-                concurrency=concurrency,
-                auto_close=False,
-                parent_run_context=resolve_parent_run_context() or data.execution.run_context,
-            )
-            captured_runtime_data = capture_target.build_runtime_data()
-            if captured_runtime_data:
-                sub_flow_execution._runtime_data.update(captured_runtime_data)
-
-            captured_resources = capture_target.build_resources()
-            if captured_resources:
-                sub_flow_execution.update_runtime_resources(captured_resources)
-            if resource_bindings:
-                self._apply_sub_flow_resource_bindings(
-                    sub_flow_execution,
-                    data.execution,
-                    resource_bindings,
-                )
-
-            stream_bridge_task = asyncio.create_task(
-                self._bridge_sub_flow_runtime_stream(
-                    sub_flow_execution,
-                    data.execution,
-                )
-            )
-            try:
-                await sub_flow_execution._async_run_start(capture_target.build_input())
-                if sub_flow_execution.is_waiting():
-                    frame_id = self._make_sub_flow_frame_id(data.execution, operator["id"])
-                    frame = {
-                        "frame_id": frame_id,
-                        "status": "waiting",
-                        "parent_execution_id": data.execution.id,
-                        "parent_operator_id": operator["id"],
-                        "child_execution_id": sub_flow_execution.id,
-                        "child_flow_name": isolated_sub_flow.name,
-                        "parent_signal": data.execution._serialize_signal(data.signal),
-                        "parent_layer_marks": data._layer_marks.copy(),
-                        "parent_value": _clone_sub_flow_value(data.value),
-                        "child_saved_state": sub_flow_execution.save(),
-                        "projected_interrupts": {},
-                        "resource_bindings": dict(resource_bindings),
-                        "resource_keys": sorted(str(key) for key in sub_flow_execution.get_runtime_resources().keys()),
-                    }
-                    projected_root_ids = await self._project_child_interrupts(
-                        parent_execution=data.execution,
-                        child_execution=sub_flow_execution,
-                        frame=frame,
-                    )
-                    root_interrupt = (
-                        data.execution.get_interrupt(projected_root_ids[0])
-                        if projected_root_ids
-                        else {
-                            "id": frame_id,
-                            "type": "sub_flow",
-                            "status": "waiting",
-                            "sub_flow_frame_id": frame_id,
-                        }
-                    )
-                    return TriggerFlowPauseSignal(root_interrupt)
-                result = await self._complete_sub_flow_frame(
-                    parent_execution=data.execution,
-                    child_execution=sub_flow_execution,
-                    frame={
-                        "frame_id": self._make_sub_flow_frame_id(data.execution, operator["id"]),
-                        "status": "running",
-                        "parent_execution_id": data.execution.id,
-                        "parent_operator_id": operator["id"],
-                        "child_execution_id": sub_flow_execution.id,
-                        "child_flow_name": isolated_sub_flow.name,
-                        "parent_signal": data.execution._serialize_signal(data.signal),
-                        "parent_layer_marks": data._layer_marks.copy(),
-                        "parent_value": _clone_sub_flow_value(data.value),
-                        "projected_interrupts": {},
-                        "resource_bindings": dict(resource_bindings),
-                        "resource_keys": sorted(str(key) for key in sub_flow_execution.get_runtime_resources().keys()),
-                    },
-                    operator=operator,
-                    normalized_write_back=normalized_write_back,
-                    write_back_bindings=write_back_bindings,
-                )
-            finally:
-                stream_bridge_task.cancel()
-                with contextlib.suppress(asyncio.CancelledError):
-                    await stream_bridge_task
-
-            return result
-
-        for signal in operator["listen_signals"]:
-            self.add_handler(
-                signal["trigger_type"],
-                signal["trigger_event"],
-                call_sub_flow,
-                id=operator["id"],
-            )
+        return self._sub_flow.compile_operator(operator, trigger_flow=trigger_flow)
 
     async def async_resume_sub_flow_frame(
         self,
@@ -1077,67 +493,12 @@ class TriggerFlowBlueprint:
         root_interrupt_id: str,
         value: Any = None,
     ):
-        frames = parent_execution._get_sub_flow_frames()
-        if frame_id not in frames:
-            raise KeyError(
-                f"Can not resume TriggerFlow sub flow frame '{ frame_id }' because it was not found."
-            )
-        frame = copy.deepcopy(frames[frame_id])
-        operator = self.definition.get_operator(str(frame["parent_operator_id"]))
-        options = operator.get("options", {})
-        _, write_back_bindings = self._compile_sub_flow_bindings(
-            options.get("write_back"),
-            mode="write_back",
+        return await self._sub_flow.async_resume_frame(
+            parent_execution,
+            frame_id,
+            root_interrupt_id,
+            value,
         )
-        normalized_write_back = options.get("write_back")
-        child_flow = self._build_sub_flow_from_operator(operator)
-        child_execution = child_flow.create_execution(
-            concurrency=options.get("concurrency"),
-            auto_close=False,
-            parent_run_context=parent_execution.run_context,
-        )
-        child_execution.load(frame["child_saved_state"])
-        self._restore_sub_flow_frame_resources(child_execution, parent_execution, frame)
-
-        child_interrupt_id = frame.get("projected_interrupts", {}).get(root_interrupt_id)
-        if not child_interrupt_id:
-            raise KeyError(
-                f"Can not resume TriggerFlow sub flow frame '{ frame_id }' because root interrupt "
-                f"'{ root_interrupt_id }' is not mapped to a child interrupt."
-            )
-
-        stream_bridge_task = asyncio.create_task(
-            self._bridge_sub_flow_runtime_stream(
-                child_execution,
-                parent_execution,
-            )
-        )
-        try:
-            await child_execution.async_continue_with(str(child_interrupt_id), value)
-            if child_execution.is_waiting():
-                projected_root_ids = await self._project_child_interrupts(
-                    parent_execution=parent_execution,
-                    child_execution=child_execution,
-                    frame=frame,
-                )
-                root_interrupt = (
-                    parent_execution.get_interrupt(projected_root_ids[0])
-                    if projected_root_ids
-                    else parent_execution.get_interrupt(root_interrupt_id)
-                )
-                return TriggerFlowPauseSignal(root_interrupt or {})
-            return await self._complete_sub_flow_frame(
-                parent_execution=parent_execution,
-                child_execution=child_execution,
-                frame=frame,
-                operator=operator,
-                normalized_write_back=normalized_write_back,
-                write_back_bindings=write_back_bindings,
-            )
-        finally:
-            stream_bridge_task.cancel()
-            with contextlib.suppress(asyncio.CancelledError):
-                await stream_bridge_task
 
     def attach_sub_flow(
         self,
@@ -1153,55 +514,18 @@ class TriggerFlowBlueprint:
         parent_group_id: str | None = None,
         parent_group_kind: str | None = None,
     ):
-        self._merge_registries_from_blueprint(trigger_flow._blue_print)
-        normalized_capture, _ = self._compile_sub_flow_bindings(capture, mode="capture")
-        normalized_write_back, _ = self._compile_sub_flow_bindings(write_back, mode="write_back")
-        sub_flow_config = trigger_flow._blue_print.definition.to_dict(name=trigger_flow.name)
-        identity_base = {
-            "listen_signals": listen_signals,
-            "name": name,
-        }
-        if name is None:
-            identity_base.update(
-                {
-                    "sub_flow_name": trigger_flow.name,
-                    "sub_flow_config": sub_flow_config,
-                    "capture": normalized_capture,
-                    "write_back": normalized_write_back,
-                    "concurrency": concurrency,
-                }
-            )
-        sub_flow_instance_id = self.make_stable_identity_digest(identity_base)
-        operator_id = f"sub-flow-{ sub_flow_instance_id }"
-        operator = self.definition.add_operator(
-            id=operator_id,
-            kind="sub_flow",
-            name=name if name is not None else trigger_flow.name,
-            listen_signals=listen_signals,
-            emit_signals=[
-                self.make_signal(
-                    "event",
-                    f"SubFlow-{ sub_flow_instance_id }-Result",
-                    role="continuation",
-                )
-            ],
-            options={
-                "sub_flow_name": trigger_flow.name,
-                "sub_flow_config": sub_flow_config,
-                "capture": normalized_capture,
-                "write_back": normalized_write_back,
-                "concurrency": concurrency,
-            },
+        return self._sub_flow.attach(
+            trigger_flow,
+            listen_signals,
+            name=name,
+            capture=capture,
+            write_back=write_back,
+            concurrency=concurrency,
             group_id=group_id,
             group_kind=group_kind,
             parent_group_id=parent_group_id,
             parent_group_kind=parent_group_kind,
         )
-        self._compile_sub_flow_operator(
-            operator,
-            trigger_flow=trigger_flow,
-        )
-        return operator
 
     def add_handler(
         self,
